+++ src/main/java/com/dalcho/adme/service/Impl/VideoServiceImpl.java
@@ -12,18 +12,24 @@ import com.dalcho.adme.repository.UserRepository;
 import com.dalcho.adme.repository.VideoRepository;
 import com.dalcho.adme.service.VideoService;
 import com.dalcho.adme.system.OSValidator;
-import com.dalcho.adme.utils.video.FfmpegUtils;
-import com.dalcho.adme.utils.video.MultipartFileUtils;
+import com.dalcho.adme.utils.video.VideoUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.io.ClassPathResource;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.transaction.Transactional;
+import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -51,6 +57,22 @@ public class VideoServiceImpl implements VideoService {
 
         user = userRepository.findByNickname(user.getNickname()).orElseThrow(UserNotFoundException::new);
 
+        String resultcontent = "";
+        InputStream inputStream = file.getInputStream();
+        BufferedReader br = new BufferedReader(new InputStreamReader(inputStream));
+        String a = br.readLine();
+//        StringBuilder builder = new StringBuilder();
+//        while (true) {
+//            String line = br.readLine();
+//            if(line == null ) break;
+//            builder.append(line);
+//        }
+//        resultcontent = builder.toString();
+
+        log.info("##-###-#-###-###-### file input Stream :"+ inputStream);
+        log.info("##-###-#-###-###-### file path :"+ a);
+
+
         String uuid = UUID.randomUUID().toString();
         String uploadPath = osValidator.checkOs();
 
@@ -59,10 +81,10 @@ public class VideoServiceImpl implements VideoService {
 
         log.info("[uploadFile] data 저장 수행");
         // Static 폴더에 파일 저장 -> 추후 Server 에 저장
-        MultipartFileUtils.saveFile(file, videoFile);
+        VideoUtils.saveFile(file, videoFile);
 
         log.info("[uploadFile] Thumbnail 생성 및 저장 수행");
-        FfmpegUtils.createThumbnail(ffmpegPath, ffprobePath, videoFile);
+        VideoUtils.createThumbnail(ffmpegPath, ffprobePath, videoFile);
 
         videoRepository.save(videoFile);
 
@@ -89,7 +111,7 @@ public class VideoServiceImpl implements VideoService {
 
         VideoFile videoFile = videoRepository.findById(id).orElseThrow(FileNotFoundException::new);
         if (!file.isEmpty()) {
-            MultipartFileUtils.deleteFile(videoFile);
+            VideoUtils.deleteFile(videoFile);
         }
 
         String uuid = UUID.randomUUID().toString();
@@ -97,9 +119,9 @@ public class VideoServiceImpl implements VideoService {
 
         videoFile.update(videoRequestDto.toEntity(uuid, uploadPath));
 
-        MultipartFileUtils.saveFile(file, videoFile);
+        VideoUtils.saveFile(file, videoFile);
 
-        FfmpegUtils.createThumbnail(ffmpegPath, ffprobePath, videoFile);
+        VideoUtils.createThumbnail(ffmpegPath, ffprobePath, videoFile);
 
         VideoResultDto videoResultDto = VideoResultDto.builder()
                 .title(videoRequestDto.getTitle())
@@ -114,7 +136,7 @@ public class VideoServiceImpl implements VideoService {
     @Transactional
     public void delete(Long id) {
         VideoFile videoFile = videoRepository.findById(id).orElseThrow(FileNotFoundException::new);
-        MultipartFileUtils.deleteFile(videoFile);
+        VideoUtils.deleteFile(videoFile);
         videoRepository.deleteById(id);
     }
