+++ src/main/java/com/dalcho/adme/utils/video/VideoUtils.java
@@ -3,6 +3,10 @@ package com.dalcho.adme.utils.video;
 import com.dalcho.adme.domain.VideoFile;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import net.bramp.ffmpeg.FFmpeg;
+import net.bramp.ffmpeg.FFmpegExecutor;
+import net.bramp.ffmpeg.FFprobe;
+import net.bramp.ffmpeg.builder.FFmpegBuilder;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
@@ -10,16 +14,12 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-
-/**
- * MultipartFile 관련 유틸리티
- */
 @Slf4j
 @RequiredArgsConstructor
-public class MultipartFileUtils {
+public class VideoUtils {
 
     public static void saveFile(MultipartFile file, VideoFile videoFile) {
-        log.info("[MultipartFileUtils] saveFile 수행");
+        log.info("[VideoUtils] saveFile 수행");
 
         Path path = Paths.get(videoFile.getUploadPath());
 
@@ -36,7 +36,7 @@ public class MultipartFileUtils {
     }
 
     public static void deleteFile(VideoFile videoFile) {
-        log.info("[MultipartFileUtils] deleteFile 수행");
+        log.info("[VideoUtils] deleteFile 수행");
 
         Path path = Paths.get(videoFile.getUploadPath() + File.separator + videoFile.getUuid() + ".mp4");
         Path thumb = Paths.get(videoFile.getUploadPath() + File.separator + "thumb_" + videoFile.getUuid() + ".jpg");
@@ -47,5 +47,24 @@ public class MultipartFileUtils {
         } catch (IOException e) {
             throw new IllegalStateException("failed to delete file. " + videoFile.getUuid(), e);
         }
+    }
+
+    public static void createThumbnail(String ffmpegPath, String ffprobePath, VideoFile videoFile) throws IOException {
+
+        log.info("[VideoUtils] createThumbnail");
+
+        FFmpeg ffmpeg = new FFmpeg(ffmpegPath);
+        FFprobe ffprobe = new FFprobe(ffprobePath);
+
+        FFmpegBuilder builder = new FFmpegBuilder()
+                .overrideOutputFiles(true) // 출력이 있는경우 덮어쓰기
+                .setInput(videoFile.getUploadPath() + videoFile.getUuid() + ".mp4") // 썸네일 생성대상 파일
+                .addExtraArgs("-ss", "00:00:01") // 썸네일 추출 시작점
+                .addOutput(videoFile.getUploadPath() + "thumb_" + videoFile.getUuid() + ".jpg") // 썸네일 파일을 저장할 위치
+                .setFrames(1) // 프레임 수
+                .done();
+
+        FFmpegExecutor executor = new FFmpegExecutor(ffmpeg, ffprobe);
+        executor.createJob(builder).run();
     }
 }
